requirements.txt
discord.py>=2.3.2
apscheduler>=3.10.4
python-dotenv>=1.0.1
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
matplotlib>=3.7.0
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
asyncio-subprocess>=0.1.5
psutil>=5.9.0
requests>=2.31.0
aiohttp>=3.8.0